Update README.md
This project is a simulation of the MasterMind board game, a code-breaking game played by two players. The game involves one player creating a secret code using a set of colored pegs, while the other player attempts to guess the code within a limited number of attempts. The game implements core concepts of C++ such as strings, loops, and functions.